Added FileStream, TimeWarpFileStream, and
TimeWarpFileStreamManager classes. Also added
a SimulationObject method to get a FileStream
object. If no FileStream object exists for the
particular file, then one is created. For a
sequential simulation, just a base FileStream
object is returned but for a timewarp simulation,
a TimeWarpFileStream object is returned which
overrides the output methods. Multiple objects
can share a FileStream also. The TimeWarpFileStreamManager
is used so that output requests can be committed and
rolled back and a timestamp can be kept.
I still need to test all of this.