Modified the way GVT calculation is started. The
manager thread will get the current time and check
to see if the duration from the start time is larger
than the gvt period. When it is, the current time
becomes the start time and the calculation is started.
I think this will be better than counting events, but
I have no idea how long the gvt period should be.